refactor: remove unused printing
